fix(core): propagate errors that come from a promise's own error handlers
Roby's Promise class allows to handle errors with #on_error. Having
error handlers defined would cause ExecutionEngine's promise
"watch" method to ignore the failed promise and remote it.

It would not account for an error handler that itself would raise. This
should not happen, but it obviously *will* happen (I should say "did"),
and ignoring *that* error is a really bad thing.

Propagate the error as a framework error.