fix(test): introduce a force-kill step in plan teardown
If the "normal" teardown fails, for instance because a task fails
to stop or is blocked, the next step - in the current flow - was
to clear the plan. That was fine as long as we were not controlling
anything external, but when we do (e.g. Syskit orogen component tests),
it leaves things running outside Syskit between tests.

Add an intermediate force-kill step where we remove dependency relations,
and forcefully kill execution agents to attempt to clean things up
anyways.