fix(app): unexpected_error? now returns false for exceptions induced by a matching exception
Localized errors in the plan usually generate other errors
(such as e.g. MissionFailed/PermanentTaskFailed).

The ExecutionExpectation harness already filters some of these out,
to avoid having to list every possible consequence of an error we
actually are testing for (e.g. the stop of a component). However,
it was only checking the first level of error - that is, whether
an error itself was related to a predicate.

This commit makes it test the exception and its origins, to avoid
too many spurious problems as soon as tests use structures more
complicated than a plain task (e.g. a composition)