chore: fix rubocop offenses reported by rubocop 1.0
Migration to 1.0 will be painful as rock has been stuck on 0.82 for a
long while. This commit fixes most rubocop offenses on 1.0. There will be
a corresponding commit on the rubocop configuration for the rest once
we are ready to switch to 1.0

This still passes on the rubocop version pinned by rock-core