feat(log): add time in PlanRebuilder's event-related arrays
PlanRebuilder keeps arrays for event-related propagation information,
but was not storing the time these propagation events happened. This
was fine so far for roby-log display, which is not listing anything,
but fails utterly to analyze the list of propagation.

Change the format of these arrays. This does not change the log format