fix(droby): race condition in Logfile::Server
There is a small window between the server's #accept and the time
where the socket is registered. This causes this particular socket
to not be closed if the server gets interrupted in the middle

Yet another reason why it's a bad idea to have exceptions getting
raised in the middle of anywhere.

Register the socket right away to avoid this problem.