fix(core): do not set Task#failure_reason to an arbitrary object in #mark_failed_to_start
In most cases, the reason would be a CodeError centered on the task
itself. However, that's not always the case. For instance, if the
task fails to start because of an execution agent failure, we end
up breaking the chain of explanations (we get Task#failure_reason
the terminal event of the deployment, without explanation about why
they are related).

This commit transform non-errors, non-localized errors and/or
localized errors from other objects into a FailedToStart exception
to be stored in failure_reason. It fixes issues in execution
expectations, where MissionFailed errors would be reported as
unexpected even if the source event was indeed expected.