feat(test): add the 'within' argument to not_emit
I somehow had the wrong idea that timeout(...) was controlling how
long we would wait to make sure that the event is not emitted.
That turns out to not be the case. Add the `within:` argument
to specify that time, and set it to 0 by default. This matches
the current behavior (raise if no event is emitted within the first
cycle of execution)