feat: add a stub mode to ExternalProcess, and a global flag to control it
ExternalProcess are actually spawning subprocesses in roby tests, which
goes against the point of the "simulation" mode, where we don't start
subprocesses / external processes.

This commit adds a stub_subprocess argument to the task, and a global
ExternalProcess.stub_in_roby_simulation_mode flag to set it automatically
during tests. This flag is false for backward compatibility reasons.