fix: remove excessive files in test file discovery
The current discovery in `all` mode would add all libs from all
dependent bundles. Libs are really the part that should be
config-independent, and testing in a given bundle should be enough
for another.

Moreover, it was also loading the test/actions/test_main.rb, also
in all bundles. This is nonsensical as Main is built by the robot
config itself.

Both functionalities were introduced before the migration to the
name-based test discovery in the local bundle, which now naturally
covers both cases.