feat: add 'example' values for required arguments
These are meant for the model-based tools to be able to call simple
actions that have required arguments, when the resulting plan is not
argument-dependent and needs to be analyzed. For instance, Syskit
mass-check predicates (e.g. assert_can_deploy) can get action state
machines and/or method actions to be called and check that the Syskit
bits are viable.

This covers the 80% of the use-cases, for the rest the calls will
have still to be done manually.