feat: display only once a repeated backtrace
When we have exceptions "in cascade", it is common for the "outer"
exceptions to share the same backtrace than the "inner" ones.

This commit makes sure we display them only once in most cases.
It does not try to optimize, but really to get rid of the simplest
common case.