fix: method missing handling in Models::ActionStateMachine must be robust to not having an action interface
The action interface model in Models::ActionStateMachine is here only
to (1) allow accessing the interface's action models by their name,
as in e.g.

    action_state_machine ... do
        s = state(some_action_def)
    end

And promoting these actions when action interfaces are subclassed.
Other use-cases for the action state machine model, as e.g. in the
new Task#action_state_machine do not need an action interface.

The implementation was already meant to be able to have no action
interface model. This commit just fixes the support within
method_missing