fix: fix a longstanding problem during log replay
Regularly, but not often, log replay would fail with

  `undefined method `failed_emissions' for #<Roby::TemplatePlan`

I finally took the time to get into it. Turns out that the Roby
logs (and therefore the engine) sometimes finalizes a task while
there are still some events pending - in this case a failed_event

What happens is that the exception leads to a finalization, and later
(during processing), some exception handler marks events as
emission_failed.

While the right fix would be to stop this behavior altogether, we
also have to deal with existing logs that show this behavior. Process
finalizations at the end of the cycle instead of immediately to
workaround. Note that this is probably what also will be done at runtime.