fix: handle some edgy behavior related to WeakRef and Set in Event source handling
Event sources are stored as WeakRef to avoid creating long event chains,
where essentially *all* plan from the beginning is being kept in memory
through event propagation.

However, Set is in the end a hash (well, a bit worse than a hash, but
bear with me), which means that in case of has collisions it needs to
call the objects that have been stored ... which in a case of a WeakRef
will raise an exception.

Change the set to an array internally, keeping a set on the public
interface to avoid changing the API.