fix: race condition in Logfile::Server between acceptation of a socket and its registration
In tests, this led to the client socket not being closed, which in turn
means that the connection is not closed. In practice, the display server
is executed in its own thread which means that the socket was closed
nonetheless.

The tests were passing if they finished the server thread before it
accepted the socket, and would fail otherwise. Moreover, there was another
race if the server thread would have time to send initial data.