fix: usage of the ROBOT pattern in test file specification
The ROBOT pattern allows to specify robot-specific paths and files
that will be resolved based on the robot name. This support was broken
by the addition of the glob support, since the globbing would remove
these files.

This commit resolves ROBOT for all arguments, including globs and
*before* globbing gets resolved. In addition, it generates an error
if a glob pattern does not match anything unless --allow-empty-globs
is set.