fix: force garbage collection on server thread
We've got a heisenbug that fails on the `after do` block claiming that
it timed out waiting for server start. However, the server shouldnt
really start, as we are mocking the thread creation, but somehow the
thread is still alive when it gets to the after do block. The idea is to
force garbage collection to run by calling server stop before after do.