fix: run all REST request handlers in the execution thread by default
The current behaviour was designed around the idea that few places would need to
interact with the execution thread in a non-thread-safe way. Reality is very
different, and bugs that come from the lack of synchronization are really hard
to track and debug.

This makes all calls run inside a ExecutionEngine#execute block. For more
fine-grained access, one can disable that behaviour and explicitly use
roby_execute, or do something like split the APIs and have two handlers, one
always synchronized and one always not.