fix: use a separate thread instead of Roby promises to interact with a REST API
The issue is not with the promise itself, but with how execution expectations
run them - that is, in a blocking way.

In this instance, the completion of the promise requires the execution engine to
run. `expect_execution` will instead block waiting for the promise to finish. It
would work in some cases, if the work is queued on the engine before it calls
`join_all_waiting_work`