feat: define the filter_out_related_errors config flag in execution expectations
To avoid overly verbose tests, the execution expectation harness filters out
"related errors". For instance, testing for `emit task.failed_event` will ignore
all errors created by this emission.

This is great for user-facing tests, but the unit test suites of Syskit and Roby
need an escape hatch. For now, I created the flag locally to use in a Syskit
test case (which had the very issue of "not seeing" a plan error) so that later
we can try turning it off for all tests in these two package's unit test suites
(I'm expecting some work)