fix: allow merging delayed arguments from state when the other arg has the same value
DelayedArgumentFromState would merge only with itself. It the other task has a
static argument with the same value than resolved with the state, it would not
merge.

This is very restrictive, and is actually problematic in Syskit when e.g.
from_conf is used in a TaskContext instance.

In that case, during network generation, Syskit itself freezes delayed arguments
before during deployment. This causes the from_conf arguments to be replaced by
their values.

Then, the deployer creates a task "from scratch" that has the from_conf.
First bug here: the newly deployed task will not have the same arguments than
the initial task.

And then it won't merge with the existing deployment, which causes
yet-another-instance to be created. In that case, the visible symptom is that
the task context gets restarted at every redeploy.