fix: change the temporal scheduler to handle relationships between schedule_as and dependency
`schedule_as` indicates that a task should be considered for scheduling only if
another task is. This is used to schedule planning tasks at the point in time
where their planned task is considered for planning, in e.g. Syskit (so, deploy
a subnet only when it is needed)

However, because Syskit also has a strong constraint that all children of a
composition will be started with the composition, we implemented that the
executable? flag for compositions returns true only if the children of the
composition are also executable.

Combined, this leads to the following pattern to fail:

   root.depends_on child
   child.depends_on abstract
   abstract.planned_by planning_task
   planning_task.schedule_as abstract

Since child is not running, abstract cannot be scheduled and won't be planned,
which leads to child remaining non-executable. Deadlock

This pull request extends the analysis capability of the temporal scheduler to
be smarter w.r.t. the relationships between schedule_as and dependencies. This
should allow to replace the `executable?` hack in Syskit with schedule_as
relationships, giving the scheduler the information it needs to handle this case
(and others) properly.

Namely:

- parent.schedule_as(child) now works. It considers temporal constraints, other
  schedule_as constraints as well as the executable flag of the child
- parent.schedule_as(child) combined with child_planning_task.schedule_as(child)
  works as well.