fix: do not register the same predicate twice in a matcher
The real bug here is the matcher API that modifies its receiver forever, that
is:

    matcher = find_tasks.executable

    # later
    matcher.executable.first

will add `executable` to `matcher` (the first one). And will do it every single
time. And then match all of them (sequentially). So, possibly, we could have
millions of these predicates being checked forever and ever.

I believe changing the API will have to be done, but will take time and will
have to be done gradually / in a conservative manner. This patch at least
alleviates the major issue without changing the matcher's (horrible) behaviour