fix: refactor Global to allow subclasses to "tune" what can resolve a non-executable task
The executable flag is essentially the only flag that can be used by things
built on top of Roby to mark a task as non-schedulable. Internally to roby
itself, the planning relation is used to resolve it, but other things might (in
Syskit, for instance, the configuration precedence relation indicates such a
relationship).

Allow subclasses of Global to easily provide such information